+++ test/mock/src/closures.js
@@ -1,14 +1,34 @@
 function functionDefineAfterReturn(v) {
 
     if (!v) {
-        foo();
-        return foo();
+        functionAfterReturn();
+        return functionAfterReturn();
     }
+
+    missed1();
+    missed2();
 
     return v;
 
-    function foo() {
+    // comments
+    function missed1() {
+        console.log('missed1');
+    }
+
+    function functionAfterReturn(a) {
+        if (a) {
+            return a;
+        }
+        covered();
         return 'Hello';
+    }
+
+    function missed2() {
+        console.log('missed2');
+    }
+
+    function covered() {
+        console.log('covered');
     }
 }