+++ test/specs/node.test.js
@@ -8,7 +8,9 @@ const {
 } = require('../mock/node/lib/app.js');
 
 // test dist with sourcemap
-const { component, branch } = require('../mock/node/dist/coverage-node.js');
+const {
+    component, branch, ignore
+} = require('../mock/node/dist/coverage-node.js');
 
 foo();
 bar();
@@ -18,5 +20,6 @@ app();
 
 component();
 branch();
+ignore();
 
 console.log = log;